requirements.txt
groq>=0.9.0
python-dotenv>=1.0.1
transformers>=4.41.0
tiktoken>=0.7.0
